Reorganize helpers for create, publish, and edit project
